requirements.txt
click<=6.7
gdapi-python
